Add Integration Test base class
