+++ src/web/EmployerFavouritesWeb/Views/Apprenticeships/Index.cshtml
@@ -11,7 +11,7 @@
                     <li class="govuk-breadcrumbs__list-item">
                         <a class="govuk-breadcrumbs__link" href="#">Home</a>
                     </li>
-                    <li class="govuk-breadcrumbs__list-item" aria-current="page">Apprentices</li>
+                    <li class="govuk-breadcrumbs__list-item" aria-current="page">Apprenticeships</li>
                 </ol>
             </div>
             <h1 class="govuk-heading-xl govuk-!-margin-top-3 govuk-!-margin-bottom-6">
