+++ src/web/EmployerFavouritesWeb/Views/Apprenticeships/TrainingProvider.cshtml
@@ -1,4 +1,4 @@
-@model TrainingProviderViewModel
+﻿@model TrainingProviderViewModel
 @section BodyHeader{
     <div class="govuk-grid-row">
         <div class="govuk-grid-column-full">
@@ -18,7 +18,13 @@
                 <div class="govuk-grid-row">
                     <div class="govuk-grid-column-two-thirds">
                         <h2 class="govuk-heading-l app-task-list__heading">@Model.ProviderName</h2>
-                        @*<details class="govuk-details">
+                        <dl class="govuk-list govuk-!-margin-top-0">
+                            <dt class="app-training-provider__details-title">Level:</dt>
+                            <dd class="app-training-provider__details-definition fav-app-level">3 (equivalent to A levels at grades A to E)</dd>
+                            <dt class="app-training-provider__details-title">Typical length:</dt>
+                            <dd class="app-training-provider__details-definition fav-app-length">18 months</dd>
+                        </dl>
+                        <details class="govuk-details">
                             <summary class="govuk-details__summary">
                                 <span class="govuk-details__summary-text">
                                     Training provider details
@@ -26,19 +32,15 @@
                             </summary>
                             <div class="govuk-details__text">
                                 <dl class="govuk-list">
-                                    <dt class="app-training-provider__details-title">Distance:</dt>
-                                    <dd class="app-training-provider__details-definition">1.1 miles away <span class="govuk-!-font-weight-bold">3</span> locations</dd>
-                                    <dt class="app-training-provider__details-title">Training options:</dt>
-                                    <dd class="app-training-provider__details-definition">day release block release at your location</dd>
+                                    <dt class="app-training-provider__details-title">Website:</dt>
+                                    <dd class="app-training-provider__details-definition">https://test.com</dd>
                                     <dt class="app-training-provider__details-title">Employer satisfaction:</dt>
                                     <dd class="app-training-provider__details-definition">86%</dd>
                                     <dt class="app-training-provider__details-title">Learner satisfaction:</dt>
                                     <dd class="app-training-provider__details-definition">91%</dd>
-                                    <dt class="app-training-provider__details-title">Achievement rate:</dt>
-                                    <dd class="app-training-provider__details-definition">65%</dd>
                                 </dl>
                             </div>
-                        </details>*@
+                        </details>
                     </div>
                     @*<div class="govuk-grid-column-one-third">
                         <a href="#" class="govuk-button app-task-list__training-provider-content">Recruit an apprentice</a>
